+++ Assets/Tests/Runtime/MediaAssetConcatenateTest.cs
@@ -1,0 +1,14 @@
+/* 
+*   VideoKit
+*   Copyright © 2025 Yusuf Olokoba. All Rights Reserved.
+*/
+
+namespace VideoKit.Tests {
+
+    using UnityEngine;
+
+    internal sealed class MediaAssetConcatenateTest : MonoBehaviour { // INCOMPLETE
+
+        
+    }
+}